feat: Register neoforge configuration
This commit migrates from self-made configuration system to using
neoforge builtin configs. This in turn minimizes the amount of errors
that could be made, makes mod configs similar to other configurations,
enables in-game editing.

To support legacy configuration - it was marked as deprecated and I
suppose we can allow 3-6 months before disabling it completely and
bumping major version.